add-future-clear

+++ packages/core_generator/lib/core_generator.dart
@@ -9,11 +9,14 @@ import 'generated/model_local_generated.dart';
 import 'generated/model_remote_generated.dart';
 import 'generated/remote_data_source_generated.dart';
 import 'generated/repository_generated.dart';
+import 'generated/service_generated.dart';
 import 'generated/use_case_generated.dart';
 import 'settings/config.dart';
 import 'utils/utils.dart';
 
 export 'utils/utils.dart';
+
+part 'g/build_model.dart';
 
 class CoreGenerator {
   CoreGenerator(this.pubspecFile);
@@ -96,16 +99,71 @@ class CoreGenerator {
       ),
     );
 
+    final cs = [
+      if (config.pubspec.config.dirLocalDataSource != null)
+        config.pubspec.config.dirLocalDataSource,
+      if (config.pubspec.config.dirLocalModel != null)
+        config.pubspec.config.dirLocalModel,
+      if (config.pubspec.config.dirModel != null)
+        config.pubspec.config.dirModel,
+      if (config.pubspec.config.dirRemoteDataSource != null)
+        config.pubspec.config.dirRemoteDataSource,
+      if (config.pubspec.config.dirRemoteModel != null)
+        config.pubspec.config.dirRemoteModel,
+      if (config.pubspec.config.dirRepository != null)
+        config.pubspec.config.dirRepository,
+      if (config.pubspec.config.dirUseCase != null)
+        config.pubspec.config.dirUseCase,
+      if (config.pubspec.config.localSchema != null)
+        config.pubspec.config.localSchema,
+      if (config.pubspec.config.webLocator != null)
+        config.pubspec.config.webLocator,
+    ];
+
     void defaultWriter(String contents, String path, {bool export = false}) {
-      final file = File(normalize(join(parent, path)));
-      if (file.existsSync()) {
+      final pFile = normalize(join(parent, path));
+      final file = File(pFile);
+      if (file.existsSync() && pFile.contains(modelName.toSnakeCase())) {
         file.deleteSync(recursive: true);
+      } else if (file.existsSync()) {
+        final s = file.readAsStringSync();
+        final m = modelName.toPascalCase();
+        final n = modelName.toSnakeCase();
+        final rex = [
+          '(?:',
+          '\\n\\n.*($n).*(;)',
+          '|',
+          '\\n.*($m).*(,)',
+          '|',
+          '\\n\\n.*TODO\\(gen\\).*\\n.*(@GET).*\\n.*($m\\(\\);)',
+          '|',
+          '\\n.*\\.\\.registerFactory\\<${m}LocalDataSource\\>.*new\\)',
+          '|',
+          "\\n\\nexport '$n\\/.*\\n.*if.*$n.*;",
+          ')',
+        ];
+
+        file.writeAsStringSync(s.replaceFirst(RegExp(rex.join()), ''));
+      }
+      for (int i = 1; i < 3; i++) {
+        final s = pFile.split('/');
+        final sv = s.sublist(0, s.length - i).join('/');
+        final absolute = Directory(sv);
+        if (!sv.contains('.dart') &&
+            sv.contains(modelName.toSnakeCase()) &&
+            !cs.contains(sv)) {
+          if (absolute.existsSync()) {
+            absolute.deleteSync(recursive: true);
+          }
+        }
       }
     }
 
     writer ??= defaultWriter;
 
-    if (!parent.contains('.dart')) {
+    if (!parent.contains('.dart') &&
+        parent.contains(modelName.toSnakeCase()) &&
+        !cs.contains(parent)) {
       final absoluteOutput = Directory(parent);
 
       if (absoluteOutput.existsSync()) {
@@ -114,41 +172,6 @@ class CoreGenerator {
     }
 
     return writer;
-  }
-
-  void buildModel(
-    String modelName, {
-    Config? config,
-    FileWriter? writer,
-  }) {
-    final initial = _build(
-      modelName,
-      pathWork: (c) => c.dirModel,
-    );
-    writer ??= initial.$1;
-
-    final reader = initial.$2;
-
-    final fileName = modelName.toSnakeCase();
-
-    //#region gen
-    writer?.call(
-      generatedModel(modelName),
-      '${fileName}_model.dart',
-    );
-    //#endregion
-
-    //#region export
-    const fileExport = 'models.dart';
-    writer?.call(
-      generatedExportModel(
-        modelName,
-        reader?.call(fileExport) ?? '',
-      ),
-      fileExport,
-      export: true,
-    );
-    //#endregion
   }
 
   void buildRemoteModel(
@@ -188,7 +211,6 @@ class CoreGenerator {
 
   void _deleteBuildRemoteModel(
     String modelName, {
-    Config? config,
     FileWriter? writer,
   }) {
     final initial = _deleteBuild(
@@ -257,6 +279,35 @@ class CoreGenerator {
     //#endregion
   }
 
+  void _deleteBuildLocalModel(
+    String modelName, {
+    Config? config,
+    FileWriter? writer,
+  }) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.dirLocalModel,
+    );
+    writer ??= initial;
+
+    final fileName = modelName.toSnakeCase();
+
+    //#region gen
+    writer?.call('', '$fileName/i_$fileName.dart');
+    writer?.call('', '$fileName/${fileName}_local.dart');
+    config ??= loadPubspecConfigOrNull(pubspecFile);
+    if (config?.pubspec.config.webLocator != null) {
+      writer?.call('', '$fileName/${fileName}_local_stub.dart');
+    }
+    //#endregion
+
+    //#region export
+    const fileExport = 'local.dart';
+    writer?.call('', fileExport, export: true);
+    _deleteBuildLocalSchema(modelName);
+    //#endregion
+  }
+
   void _buildLocalDataSourceLocator(
     String modelName, {
     FileWriter? writer,
@@ -278,6 +329,22 @@ class CoreGenerator {
       fileExport,
       export: true,
     );
+    //#endregion
+  }
+
+  void _deleteBuildLocalDataSourceLocator(
+    String modelName, {
+    FileWriter? writer,
+  }) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.webLocator,
+    );
+    writer ??= initial;
+
+    //#region export
+    const fileExport = '';
+    writer?.call('', fileExport, export: true);
     //#endregion
   }
 
@@ -303,6 +370,22 @@ class CoreGenerator {
       fileExport,
       export: true,
     );
+    //#endregion
+  }
+
+  void _deleteBuildLocalSchema(
+    String modelName, {
+    FileWriter? writer,
+  }) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.localSchema,
+    );
+    writer ??= initial;
+    //#region export
+    const fileExport = '';
+
+    writer?.call('', fileExport, export: true);
     //#endregion
   }
 
@@ -353,6 +436,37 @@ class CoreGenerator {
     //#endregion
   }
 
+  void _deleteBuildLocalDataSource(
+    String modelName, {
+    Config? config,
+    FileWriter? writer,
+  }) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.dirLocalDataSource,
+    );
+    writer ??= initial;
+
+    final fileName = modelName.toSnakeCase();
+
+    //#region gen
+    writer?.call('', '$fileName/${fileName}_local_data_source.dart');
+    config ??= loadPubspecConfigOrNull(pubspecFile);
+    if (config?.pubspec.config.webLocator != null) {
+      writer?.call('', '$fileName/${fileName}_local_data_source_stub.dart');
+    }
+    //#endregion
+
+    //#region export
+    const fileExport = 'local_data_sources.dart';
+    writer?.call('', fileExport, export: true);
+    if (config?.pubspec.config.webLocator != null) {
+      _deleteBuildLocalDataSourceLocator(modelName);
+    }
+
+    //#endregion
+  }
+
   void buildRemoteDataSource(
     String modelName, {
     Config? config,
@@ -388,6 +502,28 @@ class CoreGenerator {
     //#endregion
   }
 
+  void _deleteBuildRemoteDataSource(
+    String modelName, {
+    FileWriter? writer,
+  }) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.dirRemoteDataSource,
+    );
+    writer ??= initial;
+
+    final fileName = modelName.toSnakeCase();
+
+    //#region gen
+    writer?.call('', '$fileName/${fileName}_remote_data_source.dart');
+    //#endregion
+
+    //#region export
+    const fileExport = 'remote_data_sources.dart';
+    writer?.call('', fileExport, export: true);
+    //#endregion
+  }
+
   void buildRepository(String modelName, {Config? config, FileWriter? writer}) {
     final initial = _build(
       modelName,
@@ -413,6 +549,25 @@ class CoreGenerator {
       fileExport,
       export: true,
     );
+    //#endregion
+  }
+
+  void _deleteBuildRepository(String modelName, {FileWriter? writer}) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.dirRepository,
+    );
+    writer ??= initial;
+
+    final fileName = modelName.toSnakeCase();
+
+    //#region gen
+    writer?.call('', '$fileName/${fileName}_repository.dart');
+    //#endregion
+
+    //#region export
+    const fileExport = 'repositories.dart';
+    writer?.call('', fileExport, export: true);
     //#endregion
   }
 
@@ -444,23 +599,88 @@ class CoreGenerator {
     //#endregion
   }
 
+  void _buildApiClient(String modelName, {FileWriter? writer}) {
+    final initial = _build(
+      modelName,
+      pathWork: (c) => c.service,
+    );
+    writer ??= initial.$1;
+
+    final reader = initial.$2;
+    //#region export
+    const fileExport = '';
+
+    writer?.call(
+      generatedExportApiClient(
+        modelName,
+        reader?.call(fileExport) ?? '',
+      ),
+      fileExport,
+      export: true,
+    );
+  }
+
+  void _deleteBuildApiClient(String modelName, {FileWriter? writer}) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.service,
+    );
+    writer ??= initial;
+    const fileExport = '';
+
+    writer?.call('', fileExport, export: true);
+  }
+
+  void _deleteBuildUseCase(String modelName, {FileWriter? writer}) {
+    final initial = _deleteBuild(
+      modelName,
+      pathWork: (c) => c.dirUseCase,
+    );
+    writer ??= initial;
+
+    final fileName = modelName.toSnakeCase();
+
+    //#region gen
+    writer?.call('', '$fileName/${fileName}_use_case.dart');
+    //#endregion
+
+    //#region export
+    const fileExport = 'use_cases.dart';
+    writer?.call('', fileExport, export: true);
+    //#endregion
+  }
+
+  void buildModel(String modelName) => _buildModel(
+        modelName,
+        _build(modelName, pathWork: (c) => c.dirModel),
+      );
+
   void build(String modelName, {Config? config, FileWriter? writer}) {
     buildRemoteModel(modelName);
     buildLocalModel(modelName);
-    buildModel(modelName);
+    _buildModel(modelName, _build(modelName, pathWork: (c) => c.dirModel));
     buildLocalDataSource(modelName);
     buildRemoteDataSource(modelName);
     buildRepository(modelName);
     buildUseCase(modelName);
+    buildUseCase(modelName);
+    _buildApiClient(modelName);
   }
 
   void deleteBuild(String modelName, {Config? config, FileWriter? writer}) {
     _deleteBuildRemoteModel(modelName);
-    // buildLocalModel(modelName);
-    // buildModel(modelName);
-    // buildLocalDataSource(modelName);
-    // buildRemoteDataSource(modelName);
-    // buildRepository(modelName);
-    // buildUseCase(modelName);
+    _deleteBuildLocalModel(modelName);
+    _deleteBuildModel(
+      modelName,
+      _deleteBuild(
+        modelName,
+        pathWork: (c) => c.dirModel,
+      ),
+    );
+    _deleteBuildLocalDataSource(modelName);
+    _deleteBuildRemoteDataSource(modelName);
+    _deleteBuildRepository(modelName);
+    _deleteBuildUseCase(modelName);
+    _deleteBuildApiClient(modelName);
   }
 }
